chore: moved profile files
