+++ src/components/features/profile/index.ts
@@ -1,0 +1,4 @@
+import PrivateProfile from './private-profile';
+import Profile from './profile';
+
+export { Profile, PrivateProfile };
