+++ src/components/features/profile/tabs/Earned.tsx
