--- src/components/templates/profile/index.ts
@@ -1,4 +1,0 @@
-import PrivateProfileTemplate from './PrivateProfileTemplate';
-import ProfileTemplateLayout from './profile-template-layout';
-
-export { ProfileTemplateLayout, PrivateProfileTemplate };
