+++ src/pages/profile/[username]/issued.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 
-import { IssuedTab } from '@/components/templates/profile/tabs';
-import { ProfileTemplateLayout } from '@/components/templates/profile';
+import { Profile } from '@/components/features/profile';
+import { IssuedTab } from '@/components/features/profile/tabs';
 import { useAuth } from '@/providers/auth';
 import { gqlAnonMethods } from '@/services/hasura/api';
 import { dehydrate, QueryClient, useQuery } from '@tanstack/react-query';
@@ -53,4 +53,4 @@ export const getServerSideProps = async ({ params }) => {
     },
   };
 };
-IssuedProfile.PageLayout = ProfileTemplateLayout;
+IssuedProfile.PageLayout = Profile;
