+++ src/pages/profile/index.tsx
@@ -1,5 +1,5 @@
-import { PrivateProfileTemplate } from '@/components/templates/profile';
-import { ReceivedTab } from '@/components/templates/profile/tabs';
+import { PrivateProfile } from '@/components/features/profile';
+import { ReceivedTab } from '@/components/features/profile/tabs';
 import { useAuth } from '@/providers/auth';
 
 // TODO: make the behavior of this page better
@@ -15,4 +15,4 @@ export default function PrivateReceivedProfile() {
 
 PrivateReceivedProfile.auth = true;
 
-PrivateReceivedProfile.PageLayout = PrivateProfileTemplate;
+PrivateReceivedProfile.PageLayout = PrivateProfile;
